% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilde.R
\name{greta_assign}
\alias{greta_assign}
\title{Do some assignment with greta objects}
\usage{
greta_assign(quo_x, y)
}
\arguments{
\item{quo_x}{a rlang::quosure wrapping the first argument to \code{x}}

\item{y}{greta_array}
}
\description{
Do some assignment with greta objects
}
