% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_distrib}
\alias{has_distrib}
\title{Check whether an object has a distribution}
\usage{
has_distrib(x)
}
\arguments{
\item{an}{object}
}
\value{
length 1 logical vector
}
\description{
Check whether an object has a distribution
}
\details{
Any non greta_array object has no distribution by definition
}
