% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{in_env}
\alias{in_env}
\title{Check whether an object is in a specified environment}
\usage{
in_env(name, env)
}
\arguments{
\item{name}{an objects name}

\item{env}{an environment}
}
\value{
length 1 logical vector
}
\description{
Check whether an object is in a specified environment
}
