% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.greta_array}
\alias{is.greta_array}
\title{Check whether an object is a greta_array}
\usage{
is.greta_array(x)
}
\arguments{
\item{x}{an object}
}
\value{
length 1 logical vector
}
\description{
Check whether an object is a greta_array
}
