% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilde.R
\name{~}
\alias{~}
\title{Reworked ~ for use in greta}
\usage{
"~"(x, y)
}
\arguments{
\item{x}{an object}

\item{y}{an object}
}
\description{
Reworked ~ for use in greta
}
\details{
If y is not a greta array ~ uses rlang to construct a normal
formula call. Otherwise ~ will assign greta array, and will use
\code{`distribution<-`} if appropriate
}
