% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilde.R
\name{greta_assign}
\alias{greta_assign}
\title{Do some assignment with greta objects}
\usage{
greta_assign(expr_x, y, env)
}
\arguments{
\item{y}{greta_array}

\item{quo_x}{a rlang::quosure wrapping the first argument to \code{x}}
}
\description{
Do some assignment with greta objects
}
