% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tilde.R
\name{~}
\alias{~}
\title{Reworked ~ for pretty greta use}
\usage{
"~"(x, y)
}
\arguments{
\item{x}{object}

\item{y}{object}
}
\description{
Reworked ~ for pretty greta use
}
\details{
if y is a formula will try to do greta assignment otherwise will
convert to a normal formula
}
